p5-time-hires: add legacysupport
this port misconfigures on Tiger at least.
It thinks it finds ns time resolution (not sure why)
and  then tries to use these defines and fails.

adding legacysupport lets it build. Not sure if this
actually WORKs properly though. Better would be to see
why it is misconfiguring and fix that.

I configures properly on 10.6.8 and builds without
legacysupport there.