xorg-server; xorg-server-legacy: add Tiger version
the current version of xorg-server-legacy does not build
on Tiger, and after fixing it to build, it still crashes
so there is something there to fix.

until we get around to that, roll back to the last version
of xorg-server that did work on Tiger (Jan 2021) and copy it
over to xorg-server-legacy so that works too should someone
try to install that.